+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Removed
 - Drop support for .NET Standard 2.0 and Mono. NetVips now targets .NET 6 (`net6.0`) and .NET Framework 4.5.2 (`net452`) moving forward.
   See https://devblogs.microsoft.com/dotnet/the-future-of-net-standard/ for more information.
+- Remove deprecated utilities.
 
 ## [2.4.2] - 2024-10-13
 ### Fixed